bfast 0.7.0a: added "opt" feature support
(Lab7 ticket: BIOBUILDS-268)

+++ bfast/0.7.0a/build.sh
@@ -1,34 +1,36 @@
 #!/bin/bash
 
-command -v pkg-config >/dev/null || \
-    { echo "Could not find 'pkg-config' command" >&2; exit 1; }
-
-
 ## Configure
-[ "$BB_ARCH_FLAGS" == "<UNDEFINED>" ] && BB_ARCH_FLAGS=
-[ "$BB_OPT_FLAGS" == "<UNDEFINED>" ] && BB_OPT_FLAGS=
-[ "$BB_MAKE_JOBS" == "<UNDEFINED>" ] && BB_MAKE_JOBS=1
-CFLAGS="${CFLAGS} ${BB_ARCH_FLAGS} ${BB_OPT_FLAGS}"
+# Pull in the common BioBuilds build flags
+BUILD_ENV="${PREFIX}/share/biobuilds-build/build.env"
+if [[ ! -f "${BUILD_ENV}" ]]; then
+    echo "FATAL: Could not find build environment configuration script!" >&2
+    exit 1
+fi
+source "${BUILD_ENV}" -v
 
 # Update autoconf files for ppc64le detection
 cp -f "${RECIPE_DIR}/config.guess" "${SRC_DIR}/config/config.guess"
 cp -f "${RECIPE_DIR}/config.sub" "${SRC_DIR}/config/config.sub"
 
-# Force clang to use GNU89 inline semantics; otherwise, we'll end up with a
-# bunch of undefined symbols when building.
+# Force GNU89 inline semantics; otherwise, we'll end up with a bunch of
+# undefined symbols when building.
 if [[ $(gcc -v 2>&1 | grep -ci clang) -gt 0 ]]; then
     CFLAGS="${CFLAGS} -std=gnu89"
 fi
+if [[ $CC_IS_GNU ]]; then
+    CFLAGS="${CFLAGS} -fgnu89-inline"
+fi
 
-env CFLAGS="${CFLAGS} $(pkg-config --cflags zlib)" \
-    LDFLAGS="${LDFLAGS} $(pkg-config --libs zlib)" \
+env CC="${CC}" CFLAGS="${CFLAGS}" \
+    LD="${LD}" LDFLAGS="${LDFLAGS}" \
     ./configure --prefix="${PREFIX}" \
     --enable-intel64 --enable-largefile \
     --disable-bzlib
 
 
 ## Build
-make -j${BB_MAKE_JOBS} all
+make -j${MAKE_JOBS} all
 
 
 ## Install
